Main.CompositeIndicators.Coin — Type

Coin <: DataType

DataType that stores a set of named dictionaries.

The dictionaries are used to contruct and analyse a composite indicator model.

#Fields -input::Dict{Symbol,Any}:

-data::Dict{Symbol,Any}:

-weights::Dict{Symbol,Any}:

-log::DataFrame:

-results::Dict{Symbol,Any}:

-figures::Dict{Symbol,Any}:

Main.CompositeIndicators.ag_geomean — Method

ag_geomean(x::Vector,w::AbstractWeights)

Geometric mean of x using weights w by sqrt(prod(x .^ Int.(w)).

Main.CompositeIndicators.ag_mean — Method

ag_mean(x::Vector,w::AbstractWeights)

Arithmetic mean of x using weights w by mean(x,w).

Main.CompositeIndicators.ag_prod — Method

ag_prod(x::Vector,w::AbstractWeights)
Product of `x` using ignoring weights `w` by `prod(x)`.

Main.CompositeIndicators.aggregate! — Method

aggregate!(coin,datakey,weightskey;...)

Add the DataFrame results of composite indicator aggregation of coin.data[datakey] by coin.weights[weigtskey] to coin.data'[r_current] and coin.results["r_current"].

Arguments

-resultkey::String = "current": key that results are saved as in coin.data coin.results.

-ag_function::Function = ag_mean: function used to aggregate indicators of the same level. Functions must take agruments (x::Vector,w::AbstractWeights).

-ag_function_override::Tuple{Int64,Function} = (0,ag_mean): tuple of level and aggregation. function that is different than ag_function. Works only for a single level.

-indicators2exclude::Vector = []: indicator columns to exclude in aggregation.

-write2coin::Bool = true: if results are saved to coin.

Main.CompositeIndicators.get_meta — Method

get_meta(coin,colval_tup;column_out = names(coin.input[:indMeta]))

Return a DataFrame of a subset of the metadata.

Subset specified by a tuple of column name and a value, colval_tup::Tuple{Union{String,Symbol},Any}, and can be further defined using the column_out keyword.

Main.CompositeIndicators.get_meta — Method

get_meta(coin)

Return a DataFrame of a the metadata used to contruct the Coin.

Main.CompositeIndicators.levelnormalizeweights! — Function

levelnormalizedweights!(coin::Coin,weightkey::Symbol = :w_original)

Add a column of weights that sum to 1 for each Level to the coin.weights[weightkey] DataFrame. Resulting weights for each indicator are proportional to the number of indicators it shares a Parent indicator with and the Parent indicator weight.

Examples

julia> using DataFrames

julia> w = DataFrame(:iCode=>["L1A","L1B","L1C","L1D","L2A","L2B","L3A"],
           :Level=>[1,1,1,1,2,2,3],
           :Parent=>["L2A","L2A","L2B","L2B","L3A","L3A","NA"],
               :Weight=>[1,1,1,1,1,1,1])
7×4 DataFrame
 Row │ iCode   Level  Parent  Weight 
     │ String  Int64  String  Int64  
─────┼───────────────────────────────
   1 │ L1A         1  L2A          1
   2 │ L1B         1  L2A          1
   3 │ L1C         1  L2B          1
   4 │ L1D         1  L2B          1
   5 │ L2A         2  L3A          1
   6 │ L2B         2  L3A          1
   7 │ L3A         3  NA           1

julia> levelnormalizeweights!(w)
7×5 DataFrame
 Row │ iCode   Level  Parent  Weight  WeightNorm 
     │ String  Int64  String  Int64   Float64    
─────┼───────────────────────────────────────────
   1 │ L1A         1  L2A          1        0.25
   2 │ L1B         1  L2A          1        0.25
   3 │ L1C         1  L2B          1        0.25
   4 │ L1D         1  L2B          1        0.25
   5 │ L2A         2  L3A          1        0.5
   6 │ L2B         2  L3A          1        0.5
   7 │ L3A         3  NA           1        1.0

julia> w2 = copy(w);w2[5,:Weight] = 4;levelnormalizeweights!(w2)
7×5 DataFrame
 Row │ iCode   Level  Parent  Weight  WeightNorm 
     │ String  Int64  String  Int64   Float64    
─────┼───────────────────────────────────────────
   1 │ L1A         1  L2A          1         0.4
   2 │ L1B         1  L2A          1         0.4
   3 │ L1C         1  L2B          1         0.1
   4 │ L1D         1  L2B          1         0.1
   5 │ L2A         2  L3A          4         0.8
   6 │ L2B         2  L3A          1         0.2
   7 │ L3A         3  NA           1         1.0

Main.CompositeIndicators.new_coin — Method

new_coin(indData,indMeta)

Return a new Coin object to represent a composite indicator model.

Indicator data is defined by indData. Model structure and weighting scheme is defined by indMeta.

Main.CompositeIndicators.norm_competepercentile — Method

norm_competepercentile(x::Vector)

Return the percentile of each element of x where tied values are assigned identical results.

Uses the competerank to rank values. missing values are assigned a result of missing.

Examples

julia> norm_competepercentile([0,0,1,2,2,3])
6-element Vector{Float64}:
 0.0
 0.0
 0.4
 0.6
 0.6
 1.0

 julia> norm_competepercentile([0,0,1,2,missing,3])
6-element Vector{Union{Missing, Float64}}:
 0.0
 0.0
 0.5
 0.75
  missing
 1.0

Main.CompositeIndicators.norm_minmax — Method

norm_minmax(x::Vector)

Return min-max normalization of x.

Main.CompositeIndicators.norm_skipzeropercentile — Method

norm_skipzeropercentile(x::Vector)

Return the percentile of each nonzero element of x where tied values are assigned identical results.

Uses the competerank to rank nonzero values. '0' values are assigned a result of 0. missing values are assigned a result of missing.

#Examples jldoctest julia> norm_skipzeropercentile([0,0,1,2,2,3]) 6-element Vector{Float64}: 0.0 0.0 0.0 0.3333 0.3333 1.0 ```jldoctest julia> norm_skipzeropercentile([0,0,1,2,missing,3]) 6-element Vector{Union{Missing, Float64}}: 0.0 0.0 0.0 0.5 missing 1.0

Main.CompositeIndicators.normalize! — Method

normalize!(coin;norm_function::Function,datakey::String,normalizedkey::String)

Add a DataFrame of coin.data[datakey] normalized by norm_function to coin.data[norm_*normalizedkey].

Normalize by applying norm_function to each column of the DataFrame. norm_function must take and return a Vector. If datakey is unspecified, apply norm_function to coin.data["original"]. If normalizedkey is unspecified, add result with key norm_*String(Symbol(norm_function)).

Main.CompositeIndicators.reaggregate_excludeindicator! — Method

reaggregate_excludeindicator!(coin,logkey;indicators2exclude)

Add the DataFrame results of composite indicator aggregation of coin.data[datakey] by coin.weights[weigtskey] to coin.data'[r_current] and coin.results["r_current"].

Call the aggregate function using the arugments saved in the coin.log entry for logkey excluding indicators in indicators2exclude. Results are saved as coin.data'[Symbol("r_ex",join(indicators2exclude))]